Move exception tests to another file + ruff
