+++ test/asynccontextmanager_test.py
@@ -1,4 +1,3 @@
-import asyncio
 from contextlib import asynccontextmanager
 import pytest
