+++ test/gevent_test.py
@@ -6,5 +6,7 @@ import sys
 def test_gevent():
     # Run it in a separate process because gevent modifies a lot of modules
     fn = os.path.join(os.path.dirname(__file__), "_gevent.py")
-    ret = subprocess.run([sys.executable, fn], stdout=sys.stdout, stderr=sys.stderr, timeout=5)
+    ret = subprocess.run(
+        [sys.executable, fn], stdout=sys.stdout, stderr=sys.stderr, timeout=5
+    )
     assert ret.returncode == 0
