+++ test/translate_test.py
@@ -1,4 +1,3 @@
-
 from synchronicity import Synchronizer
